refactor: move report rendering into its own module
